moved some of the reset password code around
